refactor : 패키지 구조 변경 및 계산로직 간소화
